Clarified keywise problem text
When I was working through this I got stuck for a bit on this problem because it wasn't clear to me what to do with the "user" rows. Eventually I figured out that I should just add them with the name as the key, but the additional guidance here would have been helpful.